fix: double type 연산 문제로 생긴 버그 수정
- double 타입을 계산에 그대로 적용하면 문제가 생길 수 있음(소수점 연산에 집중, 부정확한 계산이 생길 수 있음)
- 현업에서 돈계산할 때, 정밀한 계산이 가능해야하기 때문에 BigDecimal 사용
- BigDecimal divide 할 때, 소숫점 scale 설정을하지 않으면 나눠질때 사용되는 값에 따라 scale을 설정함

+++ src/main/java/wooteco/subway/fare/domain/CalculateFareService.java
@@ -7,10 +7,9 @@ import wooteco.subway.maps.line.domain.Line;
 import wooteco.subway.members.member.domain.LoginMember;
 
 public class CalculateFareService {
-    public double calculateFare(LoginMember loginMember, List<Line> lines, int distance) {
+    public int calculateFare(LoginMember loginMember, List<Line> lines, int distance) {
         final LineFareRule lineFareRule = new LineFareRuleImpl();
-        final double fare =
-                DistanceFareType.calculateFare(distance) + lineFareRule.maxLineFare(lines);
+        final int fare = DistanceFareType.calculateFare(distance) + lineFareRule.maxLineFare(lines);
 
         return DiscountAgeType.discountByAge(loginMember, fare);
     }
