+++ src/main/java/wooteco/subway/fare/domain/LineFareRule.java
@@ -5,5 +5,5 @@ import java.util.List;
 import wooteco.subway.maps.line.domain.Line;
 
 public interface LineFareRule {
-    double maxLineFare(List<Line> lines);
+    int maxLineFare(List<Line> lines);
 }
